online voting system
Online Voting is a web-based voting system that will help you manage your elections easily and securely. This voting system can be used for casting votes during the elections held in colleges,versity etc. In this system the voter do not have to go to the polling booth to cast their vote.